add user item repository
