+++ src/packages/user/user/entity-actions/unlock/unlock-user.action.ts
@@ -1,4 +1,5 @@
-import { type UmbUnlockUserRepository, UmbUserRepository } from '../../repository/index.js';
+import { type UmbUnlockUserRepository } from '../../repository/index.js';
+import { UmbUserItemRepository } from '../../repository/item/user-item.repository.js';
 import { UmbEntityActionBase } from '@umbraco-cms/backoffice/entity-action';
 import { UmbContextConsumerController } from '@umbraco-cms/backoffice/context-api';
 import { UmbControllerHostElement } from '@umbraco-cms/backoffice/controller-api';
@@ -10,12 +11,12 @@ import {
 
 export class UmbUnlockUserEntityAction extends UmbEntityActionBase<UmbUnlockUserRepository> {
 	#modalManager?: UmbModalManagerContext;
-	#itemRepository: UmbUserRepository;
+	#itemRepository: UmbUserItemRepository;
 
 	constructor(host: UmbControllerHostElement, repositoryAlias: string, unique: string) {
 		super(host, repositoryAlias, unique);
 
-		this.#itemRepository = new UmbUserRepository(this.host);
+		this.#itemRepository = new UmbUserItemRepository(this.host);
 
 		new UmbContextConsumerController(this.host, UMB_MODAL_MANAGER_CONTEXT_TOKEN, (instance) => {
 			this.#modalManager = instance;
