+++ src/packages/user/user/repository/user.repository.ts
@@ -1,17 +1,14 @@
 import { UmbUserDetailDataSource, UmbUserSetGroupDataSource } from '../types.js';
 import { UmbUserServerDataSource } from './sources/user.server.data.js';
-import { UmbUserItemServerDataSource } from './sources/user-item.server.data.js';
-import { UmbUserItemStore } from './user-item.store.js';
 import { UmbUserSetGroupsServerDataSource } from './sources/user-set-group.server.data.js';
 
 import { UmbUserRepositoryBase } from './user-repository-base.js';
 import type { UmbControllerHostElement } from '@umbraco-cms/backoffice/controller-api';
-import { UmbDetailRepository, UmbItemDataSource, UmbItemRepository } from '@umbraco-cms/backoffice/repository';
+import { UmbDetailRepository } from '@umbraco-cms/backoffice/repository';
 import {
 	CreateUserRequestModel,
 	CreateUserResponseModel,
 	UpdateUserRequestModel,
-	UserItemResponseModel,
 	UserResponseModel,
 } from '@umbraco-cms/backoffice/backend-api';
 import { UmbNotificationContext } from '@umbraco-cms/backoffice/notification';
@@ -23,42 +20,16 @@ export type UmbUserDetailRepository = UmbDetailRepository<
 	UserResponseModel
 >;
 
-export class UmbUserRepository
-	extends UmbUserRepositoryBase
-	implements UmbUserDetailRepository, UmbItemRepository<UserItemResponseModel>
-{
+export class UmbUserRepository extends UmbUserRepositoryBase implements UmbUserDetailRepository {
 	#detailSource: UmbUserDetailDataSource;
-	#itemSource: UmbItemDataSource<UserItemResponseModel>;
-	#itemStore?: UmbUserItemStore;
 	#setUserGroupsSource: UmbUserSetGroupDataSource;
-
 	#notificationContext?: UmbNotificationContext;
 
 	constructor(host: UmbControllerHostElement) {
 		super(host);
 
 		this.#detailSource = new UmbUserServerDataSource(this.host);
-		this.#itemSource = new UmbUserItemServerDataSource(this.host);
 		this.#setUserGroupsSource = new UmbUserSetGroupsServerDataSource(this.host);
-	}
-
-	// ITEMS:
-	async requestItems(ids: Array<string>) {
-		if (!ids) throw new Error('Ids are missing');
-		await this.init;
-
-		const { data, error } = await this.#itemSource.getItems(ids);
-
-		if (data) {
-			this.#itemStore?.appendItems(data);
-		}
-
-		return { data, error, asObservable: () => this.#itemStore!.items(ids) };
-	}
-
-	async items(ids: Array<string>) {
-		await this.init;
-		return this.#itemStore!.items(ids);
 	}
 
 	// DETAILS
